- Moved logger to its own file. - Changed it from middleware to a regular function that needs to   be called by any route handlers that want to use it. It adds seven   lines to those functions, but the logger code is much easier to deal   with now. - Added that code to every almost route called by the front on the   search and comparison pages. - It finally logs useful information when errors occur - Had to make return_err_with_trace async - Moved redundant config stuff to backend/config
