+++ backend/db/config.py
@@ -1,46 +1,7 @@
 """Configuration for TermHub database"""
-import os
 from typing import Dict, List
 
-from backend.config import PROJECT_ROOT
-
-DB_DIR = os.path.dirname(os.path.realpath(__file__))
-BACKEND_DIR = os.path.join(DB_DIR, '..')
-DOCS_DIR = os.path.join(PROJECT_ROOT, 'docs')
-TERMHUB_CSETS_PATH = os.path.join(PROJECT_ROOT, 'termhub-csets')
-DATASETS_PATH = os.path.join(TERMHUB_CSETS_PATH, 'datasets', 'prepped_files')
-OBJECTS_PATH = os.path.join(TERMHUB_CSETS_PATH, 'objects')
-DDL_JINJA_PATH_PATTERN = os.path.join(DB_DIR, 'ddl-*.jinja.sql')
-
-CONFIG = {
-    'server': os.getenv('TERMHUB_DB_SERVER'),
-    'driver': os.getenv('TERMHUB_DB_DRIVER'),
-    'host': os.getenv('TERMHUB_DB_HOST'),
-    'user': os.getenv('TERMHUB_DB_USER'),
-    'db': os.getenv('TERMHUB_DB_DB'),
-    'schema': os.getenv('TERMHUB_DB_SCHEMA'),
-    'pass': os.getenv('TERMHUB_DB_PASS'),
-    'port': os.getenv('TERMHUB_DB_PORT'),
-    'personal_access_token': os.getenv('GH_LIMITED_PERSONAL_ACCESS_TOKEN')
-}
-CONFIG_LOCAL = {
-    'server': os.getenv('TERMHUB_DB_SERVER_LOCAL'),
-    'driver': os.getenv('TERMHUB_DB_DRIVER_LOCAL'),
-    'host': os.getenv('TERMHUB_DB_HOST_LOCAL'),
-    'user': os.getenv('TERMHUB_DB_USER_LOCAL'),
-    'db': os.getenv('TERMHUB_DB_DB_LOCAL'),
-    'schema': os.getenv('TERMHUB_DB_SCHEMA_LOCAL'),
-    'pass': os.getenv('TERMHUB_DB_PASS_LOCAL'),
-    'port': os.getenv('TERMHUB_DB_PORT_LOCAL'),
-    'personal_access_token': os.getenv('GH_LIMITED_PERSONAL_ACCESS_TOKEN')
-}
-
-def override_schema(schema: str):
-    CONFIG['schema'] = schema
-
-
-def get_schema_name():
-    return CONFIG['schema']
+from backend.config import CONFIG, CONFIG_LOCAL
 
 
 def invert_list_dict(d1: Dict[str, List[str]]) -> Dict[str, List[str]]:
